refactor: querynode moves into osrmtype
